added sample queries

+++ database/createtable.sql
@@ -4,7 +4,7 @@ CREATE TABLE salaries (
     team TEXT NOT NULL,
     winpercentage DECIMAL(4,3) NOT NULL,
     teamyear SMALLINT NOT NULL,
-    roundmade TINYINT NOT NULL,
+    roundmade SMALLINT NOT NULL,
     positiongroup TEXT NOT NULL,
     payrollamount INT NOT NULL,
     teamyearID TEXT NOT NULL
